refactor #504 textView
