+++ public/locales/he/translation.json
@@ -47,15 +47,15 @@
     "on": "ב",
     "occurred": "התרחשו",
     "accidents": "תאונות",
-    "fatal": {
+    "severity_fatal_count": {
       "singular": "קטלנית",
       "plural": "קטלניות"
     },
-    "severe": {
+    "severity_severe_count": {
       "singular": "קשה",
       "plural": "קשות"
     },
-    "light": {
+    "severity_light_count": {
       "singular": "קלה",
       "plural": "קלות"
     }
